backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
openai==1.57.4
python-dotenv==1.0.1
reportlab==4.2.5
Pillow==11.1.0
httpx==0.28.1
requests==2.32.3
python-multipart==0.0.20
